Implement nonRNAVFirstFix Fixes #230
