+++ server/src/controllers/verifiers/hasValidFirstFix.mts
@@ -42,11 +42,19 @@ export default async function hasValidFirstFix({
     // Finally have all the info necessary to verify the first fix
     else if (!SIDInformation.Fixes.includes(firstFix)) {
       result.data.status = "Error";
-      result.data.message = `First fix ${firstFix} is not in the list of fixes for ${SID}: ${SIDInformation.Fixes.join(
-        ", "
-      )}`;
-      result.data.messageId = "firstFixNotInSID";
       result.data.priority = 1;
+      // The message sent is different depending on whether it's an RNAV SID.
+      if (SIDInformation.IsRNAV) {
+        result.data.message = `First fix ${firstFix} is not in the list of fixes for ${SID}: ${SIDInformation.Fixes.join(
+          ", "
+        )}`;
+        result.data.messageId = "firstFixNotInRNAVSID";
+      } else {
+        result.data.message = `First fix ${firstFix} is not in the list of fixes for ${SID}: ${SIDInformation.Fixes.join(
+          ", "
+        )}, or requires coordination with the online departure controller.`;
+        result.data.messageId = "firstFixNotInSID";
+      }
     } else {
       result.data.status = "Information";
       result.data.message = `First fix ${firstFix} is valid for ${SID}.`;
